Remove resuming multipart uploads (#741)
ListParts is never fully complete in its output, there is always a
possibility that the List is partial where another
PutObjectPart() is being uploaded in parallel.

While it may seem like ListParts and ListMultipartUploads
provide the possibility of resuming an upload, it would
be a mistake as these APIs were inherently never meant do
these things. This is one of the reasons why Amazon
recommends that one should avoid using these API calls to
complete or add parts to an upload - they instead
recommend to remember the parts being uploaded at the client.