Abort failed multipart uploads on RuntimeException (#786)
If a file > 5MB is being uploaded to a bucket, this client will initiate
a multipart upload.  However, if in the course of uploading that file a
RuntimeException occurs, this library will not abort the multipart
upload.  This is especially possible when streaming data across a
network connection into a bucket.  This was introduced in #741 when
resuming failed multipart uploads was removed.

Reproduction steps:

* Use this script: https://gist.github.com/Dorthu/b75f2a4ad9e0e453a040412438d3fe32
* Replace the server URL and port, access key, secret key, and bucket name with
  valid values
* Compile/run the script
* Observe the state of multipart uploads in the bucket after running
  with `s3cmd multipart s3://bucket-name` (or another tool)

Subsequent runs of the script create new multipart uploads in the
bucket.

I specifically noticed this bug when using this library against Ceph, if
that's relevant.