Refactor builder framework (#929)
One issue with the regular implementation of the builder pattern is the
duplication of same variables across the actual class being built (e.g.
MakeBucketArgs) and the corresponding builder (MakeBucketArgs.Builder)

This can be avoided by making use of Java's `Consumer` functional
interface. Instead of keeping all the builder data in its instance
variables, we can rather keep lambda expressions for setting the data in
the object being built. Coupled with some use of generics, the logic of
executing all the stored lambda expressions can be then pushed to a
parent class.

Benefits:
 - No need to add constructors in both args and builder classes.
 - No need to define any instance variables in any of the concrete
 builder classes

New class hierarchy:

 - BaseArgs (abstract base class - no common logic as of now)
   - BucketBaseArgs (abstract - bucket specific common logic)
     - MakeBucketArgs (concrete implementation of make-bucket args)
 - BaseArgs.Builder (common logic related to building args)
   - BucketArgs.Builder (fluent setters for bucket common data)
     - MakeBucketArgs.Builder (fluent method for objectLock arg and the
     simple one-line `build()` method

With this approach, the only boilerplate code in concrete args classes
will be:
 - one-line `builder()` method inside the concrete args class
 - one-line `build()`  method inside the concrete builder class