Implement Args builder for listObjects API
Added the new args class `ListObjectsArgs` to be used with the
listObjects API. Example usage:

```
Iterable<Result<Item>> results = minioClient.listObjects(
  ListObjectsArgs.builder().
    bucket("my-bucketname").
    includeUserMetadata(true).
    startAfter("start-after-entry").
    prefix("my-obj").
    maxKeys(100).
    fetchOwner(true)
  );
```

Also refactored some of the existing code to reduce duplication of code:

- Extract common part of the result of both V1 and V2 API into
`ListBucketResult`, and made the two specific classes
(`ListBucketResultV1` and `ListBucketResultV2`) extend from it. The main
difference between the two is the naming of the continuation token
(`continuationToken` in V2 vs `marker` in V1). Since this is only a
naming difference, both results can be treated the same way by
abstracting out the methods related to these result fields.

- Similarly, most of the logic of the iterators for the response of V1
and V2 was also common. Extracted it out into a new abstract class
called `ObjectIterator`.