Few enhancements in the builder framework (#941)
- Infer the type of the Args class by looking it up through reflection
(`getEnclosingClass()`), so that there is no need for concrete builder
classes to implement the `build()` method.

- Change the generic type notations `<T, B>` to `<B, A>` to make them
more intuitive (`B -> Builder` and `A -> Args`)

- When inheriting from a parent builder class in an abstract builder
class like `BucketArgs.Builder`, pass the type parameters of the current
class (`<B, A>`) as is. This ensures that the builder methods,
irrespective of what order they're called in, will always return an
object with type of the original builder class.