Implement Args builder for getObject API
New Argument class: `GetObjectArgs`

Arguments supported:
 - offset (Long)
 - length (Long)
 - ssec (ServerSideEncryptionWithCustomerKey)

Since there are multiple APIs / args classes that need sse as an
argument, introduced an intermediate abstract class `SsecObjectArgs`
that supports the `ssec` argument, and provides common logic related to
it e.g. validations.

Since the return type of the methods that support downloading the
object as a file is `void`, we cannot name it same as `getObject`. So
introduced a new method called `downloadObject` for downloading the
object, which internally uses `getObject`. And created a separate args
class `DownloadObjectArgs` for it.

Also,

- Enhanced the method `FunctionalTest#handleException` to support
logging of the arguments as well.
- Moved the class `ServerSideEncryptionWithCustomerKey` outside of
`ServerSideEncryption` and made it into a standalone class.